requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.3.5
pycryptodome==3.9.7
requests==2.22.0
selenium==3.141.0
soupsieve==2.0
urllib3==1.25.3
